renamed file and cleaned up init
