+++ application/jobs/minimal_training_pytorch_cnn/app/config/config_fed_client.conf
@@ -81,7 +81,7 @@
     path = "nvflare.app_opt.pt.file_model_persistor.PTFileModelPersistor"
     args {
       model {
-        path = "models.mini_model.MiniCNNForTesting"
+        path = "models.models_for_testing.MiniCNNForTesting"
       }
     }
   }
