+++ application/jobs/minimal_training_pytorch_cnn/app/custom/models/models_for_testing.py
